refactor: relocated InstanceId and InstanceInfo
