+++ src/Application/Domain/Model/Instance/InstanceInfo.cs
@@ -1,6 +1,6 @@
 ﻿using ScriptBee.Domain.Model.Project;
 
-namespace ScriptBee.Domain.Model.Analysis;
+namespace ScriptBee.Domain.Model.Instance;
 
 public record InstanceInfo(
     InstanceId Id,
