+++ src/Application/Domain/Service.Project/Analysis/GetCurrentInstanceService.cs
@@ -1,4 +1,4 @@
-﻿using ScriptBee.Domain.Model.Analysis;
+﻿using ScriptBee.Domain.Model.Instance;
 using ScriptBee.Domain.Model.Project;
 using ScriptBee.UseCases.Project.Analysis;
