+++ src/Application/Domain/Service.Project/Analysis/TriggerAnalysisService.cs
@@ -1,5 +1,6 @@
 ﻿using ScriptBee.Common;
 using ScriptBee.Domain.Model.Analysis;
+using ScriptBee.Domain.Model.Instance;
 using ScriptBee.Domain.Model.Project;
 using ScriptBee.Domain.Model.ProjectStructure;
 using ScriptBee.Ports.Instance;
