+++ src/Application/Ports/Driven/Ports.Instance/ICreateProjectInstance.cs
@@ -1,4 +1,4 @@
-﻿using ScriptBee.Domain.Model.Analysis;
+﻿using ScriptBee.Domain.Model.Instance;
 using ScriptBee.Domain.Model.Project;
 
 namespace ScriptBee.Ports.Instance;
