+++ src/Application/Ports/Driven/Ports.Plugins/IGetPlugins.cs
@@ -1,4 +1,4 @@
-﻿using ScriptBee.Domain.Model.Analysis;
+﻿using ScriptBee.Domain.Model.Instance;
 using ScriptBee.Domain.Model.Plugin;
 
 namespace ScriptBee.Ports.Plugins;
