+++ src/Application/Ports/Driven/Ports.Plugins/IGetScriptLanguages.cs
@@ -1,5 +1,5 @@
 ﻿using OneOf;
-using ScriptBee.Domain.Model.Analysis;
+using ScriptBee.Domain.Model.Instance;
 using ScriptBee.Domain.Model.ProjectStructure;
 
 namespace ScriptBee.Ports.Plugins;
