+++ src/Application/Ports/Driving/UseCases.Analysis/ICalculationInstanceUseCase.cs
@@ -1,4 +1,4 @@
-﻿using ScriptBee.Domain.Model.Analysis;
+﻿using ScriptBee.Domain.Model.Instance;
 
 namespace ScriptBee.UseCases.Analysis;
