+++ src/Application/Ports/Driving/UseCases.Project/Analysis/IGetCurrentInstanceUseCase.cs
@@ -1,4 +1,4 @@
-﻿using ScriptBee.Domain.Model.Analysis;
+﻿using ScriptBee.Domain.Model.Instance;
 using ScriptBee.Domain.Model.Project;
 
 namespace ScriptBee.UseCases.Project.Analysis;
