+++ src/Application/Ports/Driving/UseCases.Project/Context/ClearContextCommand.cs
@@ -1,4 +1,4 @@
-﻿using ScriptBee.Domain.Model.Analysis;
+﻿using ScriptBee.Domain.Model.Instance;
 using ScriptBee.Domain.Model.Project;
 
 namespace ScriptBee.UseCases.Project.Context;
