+++ test/Adapters/Driven/Rest.Tests/GetScriptLanguagesAdapterTest.cs
@@ -1,5 +1,5 @@
 ﻿using NSubstitute;
-using ScriptBee.Domain.Model.Analysis;
+using ScriptBee.Domain.Model.Instance;
 using ScriptBee.Domain.Model.Plugin;
 using ScriptBee.Domain.Model.Plugin.Manifest;
 using ScriptBee.Domain.Model.Project;
