+++ test/Adapters/Driving/Web.Tests/EndpointDefinitions/Context/GetLinkersEndpointTest.cs
@@ -1,8 +1,8 @@
 ﻿using System.Net;
 using Microsoft.Extensions.DependencyInjection;
 using NSubstitute;
-using ScriptBee.Domain.Model.Analysis;
 using ScriptBee.Domain.Model.Context;
+using ScriptBee.Domain.Model.Instance;
 using ScriptBee.Domain.Model.Project;
 using ScriptBee.Tests.Common;
 using ScriptBee.UseCases.Project.Context;
