+++ test/Application/Domain/Service.Project.Tests/Analysis/GetProjectInstancesServiceTest.cs
@@ -1,5 +1,5 @@
 ﻿using NSubstitute;
-using ScriptBee.Domain.Model.Analysis;
+using ScriptBee.Domain.Model.Instance;
 using ScriptBee.Domain.Model.Project;
 using ScriptBee.Ports.Instance;
 using ScriptBee.Service.Project.Analysis;
