+++ test/Application/Domain/Service.Project.Tests/Analysis/TriggerAnalysisServiceTest.cs
@@ -1,6 +1,7 @@
 ﻿using NSubstitute;
 using ScriptBee.Common;
 using ScriptBee.Domain.Model.Analysis;
+using ScriptBee.Domain.Model.Instance;
 using ScriptBee.Domain.Model.Project;
 using ScriptBee.Ports.Instance;
 using ScriptBee.Service.Project.Analysis;
