+++ test/Application/Domain/Service.Project.Tests/ProjectStructure/CreateScriptServiceTest.cs
@@ -2,7 +2,7 @@
 using NSubstitute;
 using OneOf;
 using ScriptBee.Common;
-using ScriptBee.Domain.Model.Analysis;
+using ScriptBee.Domain.Model.Instance;
 using ScriptBee.Domain.Model.Project;
 using ScriptBee.Domain.Model.ProjectStructure;
 using ScriptBee.Ports.Files;
